AYS-152 | The Auth package has been rearranged according to endpoints.
