+++ src/test/resources/testsuite/InstitutionTestSuite.xml
@@ -3,14 +3,6 @@
 <suite name="All Test Suite">
     <test verbose="2" preserve-order="true" name="Institution Test Suite">
         <classes>
-            <class name="org.ays.tests.auth.InstitutionAuthServiceTest">
-                <methods>
-                    <include name="getTokenForValidAdmin"/>
-                    <include name="getTokenForInvalidAdmin"/>
-                    <include name="adminTokenRefresh"/>
-                    <include name="adminInvalidateToken"/>
-                </methods>
-            </class>
             <class name="org.ays.tests.institution.adminmanagementservice.PostAdminsTest">
                 <methods>
                     <include name="listAdminsWithValidPageAndPageSize"/>
