+++ src/test/resources/testsuite/SmokeSuite.xml
@@ -3,20 +3,6 @@
 <suite name="All Test Suite">
     <test name="Smoke Test">
         <classes>
-            <class name="org.ays.tests.auth.InstitutionAuthServiceTest">
-                <methods>
-                    <include name="getTokenForValidAdmin"/>
-                    <include name="adminTokenRefresh"/>
-                    <include name="adminInvalidateToken"/>
-                </methods>
-            </class>
-            <class name="org.ays.tests.auth.UserAuthServiceTest">
-                <methods>
-                    <include name="getTokenForValidUser"/>
-                    <include name="userTokenRefresh"/>
-                    <include name="userInvalidateToken"/>
-                </methods>
-            </class>
             <class name="org.ays.tests.institution.adminmanagementservice.PostAdminsTest">
                 <methods>
                     <include name="listAdminsWithValidPageAndPageSize"/>
